requirements.txt
certifi
websockets
simpleaudio
python-dotenv
openai
numpy
vosk
sounddevice==0.4.6
fastapi==0.110.2
uvicorn==0.29.0
requests==2.32.3